Do not symlink non-existing files

+++ etc/symlink-dotfiles.sh
@@ -14,9 +14,11 @@ fi
 link() {
   from="$1"
   to="$2"
-  echo "Linking '$from' to '$to'"
-  rm -f "$to"
-  ln -s "$from" "$to"
+  if [ -f "$to" ]; then
+    echo "Linking '$from' to '$to'"
+    rm -f "$to"
+    ln -s "$from" "$to"
+  fi
 }
 
 for location in $(find home -name '.*'); do
